refactor run_model and add test
